cmd/operator-sdk/{exec-entrypoint -> execentrypoint}
