+++ cmd/operator-sdk/cli/cli.go
@@ -26,7 +26,7 @@ import (
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/bundle"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/cleanup"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/completion"
-	execentrypoint "github.com/operator-framework/operator-sdk/cmd/operator-sdk/exec-entrypoint"
+	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/execentrypoint"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/generate"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/migrate"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/new"
