+++ cmd/operator-sdk/execentrypoint/ansible.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package run
+package execentrypoint
 
 import (
 	"github.com/operator-framework/operator-sdk/pkg/ansible"
